file saving code refactor

+++ server/routes/order/file/fileHandle.coffee
@@ -15,35 +15,24 @@ transformReq = (info) ->
 tryClean = (layer) ->
   (req, res, next) ->
     file = req.gerbers[layer]
-    if not file? then next()
-    else fs.unlink file, (err) ->
-      if err then next err
-      else next()
+    if not file? then next() else fs.unlink file, (err) ->
+      if err then next err else next()
 
 module.exports =
 
   order: post: [
       multerConfig.order().any()
-
       transformReq "filename"
-
       (req, res) -> query res, files: req.gerbers
     ]
 
   preview: post: [
     multerConfig.preview().any()
-
     transformReq "path"
-
     GerberToSVGMiddleware "top"
-
     GerberToSVGMiddleware "bottom"
-
     tryClean "top"
-
     tryClean "bottom"
-
     tryClean "outline"
-
     GerberToSVGMiddleware "send"
   ]
